Change Select component styles
